GameManager: New Signals and Fab.Input Rejection
Now that GameManager will take control of the input,
we need to introduce Fabrication alarms which can fire
at some random point during the input window.
This can override the decision-making, if there is a
chance that fabricated input fires this round.

Secondly the input introduces new signals, which can
be used to follow the input window and intertrial
interval timers.